src/third_party/blink/web_tests/external/wpt/web-nfc/NFCWriter_push.https-expected.txt

This is a testharness.js-based test.
PASS Test that promise is rejected with TypeError if NDEFMessageSource is invalid.
PASS 'Test that promise is rejected with SyntaxError if NDEFMessageSource contains invalid records.
FAIL NFCWriter.push should fail if abort push request before push happends. assert_throws: function "function() { throw e }" threw object "NotReadableError: No NFC adapter or cannot establish connection." that is not a DOMException AbortError: property "code" is equal to 0, expected 20
PASS NFCWriter.push should fail if signal's aborted flag is set.
PASS NFCWriter.push should fail if signal is not an AbortSignal.
PASS NFCWriter.push should fail with TypeError when invalid timeout is provided.
PASS NFCWriter.push should fail with TypeError when invalid negative timeout value is provided.
FAIL NFCWriter.push should fail with TimeoutError when timer expires. assert_throws: function "function() { throw e }" threw object "NotReadableError: No NFC adapter or cannot establish connection." that is not a DOMException TimeoutError: property "code" is equal to 0, expected 23
PASS Reject promise with NotSupportedError if NFC message size exceeds 32KB.
PASS Reject promise with SyntaxError if WebNFC Id cannot be created from provided URL.
PASS Reject promise with SyntaxError if 'json' record cannot be serialized.
PASS NFCWriter.push should fail with TypeError when invalid target value is provided.
PASS Test that WebNFC API is not accessible from iframe context.
Harness: the test ran to completion.
